Fix Directory

+++ server/subsonic/browsing.go
@@ -285,7 +285,7 @@ func (api *Router) GetArtistInfo2(r *http.Request) (*responses.Subsonic, error) 
 		similar.Name = s.Name
 		similar.AlbumCount = s.AlbumCount
 		similar.Starred = s.Starred
-		similar.UserRating = int(s.UserRating)
+		similar.UserRating = s.UserRating
 		similar.CoverArt = s.CoverArt
 		similar.ArtistImageUrl = s.ArtistImageUrl
 		response.ArtistInfo2.SimilarArtist = append(response.ArtistInfo2.SimilarArtist, similar)
@@ -354,8 +354,8 @@ func (api *Router) buildArtistDirectory(ctx context.Context, artist *model.Artis
 	if artist.PlayCount > 0 {
 		dir.Played = &artist.PlayDate
 	}
-	dir.AlbumCount = artist.AlbumCount
-	dir.UserRating = artist.Rating
+	dir.AlbumCount = int32(artist.AlbumCount)
+	dir.UserRating = int32(artist.Rating)
 	if artist.Starred {
 		dir.Starred = &artist.StarredAt
 	}
@@ -392,8 +392,8 @@ func (api *Router) buildAlbumDirectory(ctx context.Context, album *model.Album) 
 	if album.PlayCount > 0 {
 		dir.Played = &album.PlayDate
 	}
-	dir.UserRating = album.Rating
-	dir.SongCount = album.SongCount
+	dir.UserRating = int32(album.Rating)
+	dir.SongCount = int32(album.SongCount)
 	dir.CoverArt = album.CoverArtID().String()
 	if album.Starred {
 		dir.Starred = &album.StarredAt
